Fix linking of the sysrepo plugin
One "cannot" really mix object files built for static and shared
libraries. The easiest option is therefore to use dynamic linking for
the INI file reader (and to install it as a shared object properly).

Alternatively, one could get rid of the standalone `stest` and simply
build that one plugin from two .c and .cpp files, of course.